# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My Project API Documentation",
        "description": "This is a documentation for my project APIs",
        "contact": {
            "email": "admin@myproject.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/buildings/{id}/clean-reports": {
            "put": {
                "tags": [
                    "Buildings"
                ],
                "summary": "Update building and clean reports",
                "description": "Updates building level of damage and material origin, then deletes associated damage reports with report number 1",
                "operationId": "updateAndCleanReports",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Building ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "level_of_damage": {
                                        "description": "مستوى الضرر",
                                        "type": "integer",
                                        "enum": [
                                            0,
                                            1,
                                            2,
                                            3,
                                            4
                                        ]
                                    },
                                    "is_materials_from_the_neighborhood": {
                                        "description": "هل المواد من الحي",
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "تم تعديل البناية وحذف التقارير المرتبطة بها بنجاح"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Building not found"
                    }
                }
            }
        }
    },
...
}